added: truststore provider to ssl cert parser
